Fix: Dialog picker (#499)
* Added page input dialog. Scale fix.

* Fixed applying same params for report

+++ app/src/main/java/com/jaspersoft/android/jaspermobile/activities/report/ReportCastActivity.java
@@ -91,7 +91,7 @@ public class ReportCastActivity extends BaseReportActivity implements ResourcePr
     }
 
     @Override
-    protected float provideScale() {
+    protected double provideScale() {
         // To make casting device-dpi independent we are considering device density
         return 0.66f / getResources().getDisplayMetrics().density;
     }
